requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
pydantic-settings
pymongo[srv]